+++ Sources/DOR.h
@@ -33,7 +33,7 @@ typedef struct ChannelThreadData
   OS_ECB* semaphore;
   uint8_t channelNb;
   float irms;
-  int16_t samples[16];
+  float samples[16];
   int16_t sample;
   bool timerStatus;
   uint32_t currentTimeCount;
